reconfigure am role assignment prep
