+++ lib-consumer/src/test/resources/cftlib-am-role-assignments.json
